+++ users/client/profile/profile.js
@@ -1,33 +1,5 @@
-AutoForm.hooks({
-  updateProfile: {
-    before: {
-      update (user) {
-        const userUnset = user.$unset;
-        // Check if username is empty
-        if (userUnset.username === '') {
-          // Inform user about error
-          const errorMessage = TAPi18n.__('profile_usernameInvalid');
-          sAlert.error(errorMessage);
-          // Cancel form
-          return false;
-        } else {
-          // Otherwise return changes
-          return user;
-        }
-      },
-    },
-    onSuccess (operation, result, template) {
-      // Get update success message translation
-      const message = TAPi18n.__('profile_updatedSuccess');
-
-      // Alert user of success
-      sAlert.success(message);
-    },
-    onError () {
-      this.addStickyValidationError('username', 'usernameTaken');
-    },
-  },
-});
+import { Meteor } from 'meteor/meteor';
+import { Template } from 'meteor/templating';
 
 Template.profile.helpers({
   currentUser () {
